requirements.txt
flask
gunicorn
line-bot-sdk
openai
python-dotenv